// src/api/bindings/GameCreateResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GameStateView } from "./GameStateView";
import type { PlayerInfo } from "./PlayerInfo";
import type { TimeControl } from "./TimeControl";

export type GameCreateResponse = { state: GameStateView, white_player: PlayerInfo, black_player: PlayerInfo, time_control: TimeControl, };